Test ability of OTT to handle Hessian of reg_ot_cost w.r.t. weights or locations.
Computing second order derivatives induces some changes in the way we handle reg_ot_cost.

* When using backprop, everything should be handled automatically, and that's fine. This is checked with tests in this CL.

* When using implicit differentiation, computing a Hessian trigers automatic differentiation of the custom_vjp step. While feasible, a few of the numerical tricks that were working then (notably Danskin's approximation, a stop_gradient for argmax'es, but other stabilization tricks designed to stabilize computations such as recentering) were causing several errors. For instance, Danskin's approx, while beneficial and valid when computing 1st order diffs, is problematic because it "kills" meaningful second order information (Danskin's theorem stems from the fact that one term can be neglected because the gradient of the objective function w.r.t. argmax is 0. Unfortunately that does not hold in the 2nd order).

To correct this,
- a flag is proposed to control whether Danskin approx (stop_gradient) is used, independently of implicit_differentiation. By default the flag is equal to implicit_differentiation but allows overriding this option. In particular, whenever computing second order information, that flag must be set to False. Doing this automatically remains a TODO.

- A lot of computational tricks designed to stabilize computations in the forward pass of the custom_vjp are no longer valid when running that custom_vjp backwards. These have been removed.

To illustrate these findings, a series of tests checks whether the Hessian can be computed, in all regimes (autodiff/implicit, w.r.t. weights or locations).

Additionally, run tests with settings where n<m and n>m to check that implicit differentiation works in both cases.

Note finally that these tests for the Hessian matrix are flaky in the sense that they rely on a regularization strength for the implicit linear system. These regularizers need adjusting. It is therefore now possible to pass these regularization strengths as parameters when calling sinkhorn.

PiperOrigin-RevId: 372935488