add various options to set epsilon in a more clever range.
when no epsilon is passed, default behaviour is to compute average cost contained in cost_matrix (also works for grid and pointcloud) and use a 1/20 th  of that value as regularizer.

when epsilon is passed, then Geometry uses that value directly, unless user specifies a relative_epsilon=True parameter, in which case epsilon is understood as a fraction of the mean value observed in cost matrix.

in sinkhorn divergence, this behaviour could lead to possibly 3 different epsilon values if none is specified. To account for this introduction of a flag, True by default, that ensures the same epsilon scheduler is applied 3 times.

issues raised by this CL:
- handling of the case where epsilon is itself a scheduler might be mishandled in corner cases.
- In GW in particular, there might be issues in ensuring more than just the epsilon value (i.e. geometric decay parameters) are passed on.

PiperOrigin-RevId: 380205975